move to_excel and _repr_html_ back
